+++ part2/phonebook/src/components/Display.jsx
@@ -1,0 +1,13 @@
+const DisplayName=({filtered})=>{
+
+    return(
+      <div>
+      {filtered.map(person=>
+        <p key={person.name}>
+          {person.name} {person.phonenumber}</p>  
+      )}
+      </div>
+    )
+  }
+
+  export default DisplayName
