+++ part2/phonebook/src/components/Filter.jsx
@@ -1,0 +1,9 @@
+const Filter=({name,handler})=>{
+    return(
+      <div>
+        filter : <input value={name} onChange={handler}/>
+      </div>
+    )
+  
+  }
+  export default Filter
